Clothing Store WearhouseDatabase
Development of a system for inventory management and accounting for a clothing store. I have created a comprehensive table for all the items available in the store, incorporating five columns to accommodate the necessary data. The inventory comprises a minimum of 15 items, and I have utilized select statements to arrange the items based on price and present at least one statistical detail regarding them.